[#30]refactor: springParser util 패키지로 이동
