+++ BE/src/main/java/com/team12/sidedish/util/StringParser.java
@@ -1,4 +1,4 @@
-package com.team12.sidedish.service;
+package com.team12.sidedish.util;
 
 import java.util.Arrays;
 import java.util.List;
